# Compact code by using a one line for loop in CheckCells