BENCHMARKS:
The following benchmarks were done using the time command with the V3_ENCODE function call commented out

BEFORE:
```
LB1 n n 0123456789abcdefx0123456789abcdefx000b3b3
python3 ./encode.py  0.02s user 0.01s system 90% cpu 0.037 total
```
AFTER:
```
LB1 n n 0123456789abcdefx0123456789abcdefx000b3b3
python3 ./encode.py  0.02s user 0.01s system 90% cpu 0.035 total
```

As you can see there is not much of a difference, so honestly this is just more compact.